requirements.txt
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.1
django-ckeditor-5==0.2.11
django-js-asset==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.4
